restore password and settings

+++ backend/src/main/java/com/akveo/bundlejava/authentication/resetpassword/RestorePasswordService.java
@@ -6,13 +6,17 @@
 
 package com.akveo.bundlejava.authentication.resetpassword;
 
+import com.akveo.bundlejava.authentication.exception.PasswordsDontMatchException;
 import com.akveo.bundlejava.authentication.resetpassword.exception.TokenNotFoundOrExpiredHttpException;
 import com.akveo.bundlejava.user.ChangePasswordRequest;
 import com.akveo.bundlejava.user.UserService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.Base64;
 import java.util.Objects;
 
 @Service
@@ -28,7 +32,7 @@ public class RestorePasswordService {
         this.userService = userService;
     }
 
-    public void restorePassword(RestorePasswordDTO restorePasswordDTO) {
+   /* public void restorePassword(RestorePasswordDTO restorePasswordDTO) {
         RestorePassword restorePassword =
                 restorePasswordTokenRepository.findByToken(restorePasswordDTO.getToken());
 
@@ -36,6 +40,23 @@ public class RestorePasswordService {
             throw new TokenNotFoundOrExpiredHttpException();
         }
 
+        changePassword(restorePasswordDTO, restorePassword);
+        restorePasswordTokenRepository.delete(restorePassword);
+    }*/
+
+    public void restorePassword(RestorePasswordDTO restorePasswordDTO) throws IOException {
+        if (!restorePasswordDTO.getNewPassword().equals(restorePasswordDTO.getConfirmPassword())) {
+            throw new PasswordsDontMatchException();}
+        byte[] decodedBytes = Base64.getDecoder().decode(restorePasswordDTO.getToken());
+        String decodedToken = new String(decodedBytes);
+        System.out.println("decoded Token" + decodedToken + "\n");
+        ObjectMapper objectMapper = new ObjectMapper();
+        RestorePasswordTokenDto restorePasswordTokenDto = objectMapper.readValue(decodedToken , RestorePasswordTokenDto.class);
+        RestorePassword restorePassword = restorePasswordTokenRepository.findByToken(restorePasswordTokenDto.getToken());
+        System.out.println("restore Password" + restorePassword);
+        if (Objects.isNull(restorePassword) || restorePassword.isExpired()) {
+            throw new TokenNotFoundOrExpiredHttpException();
+        }
         changePassword(restorePasswordDTO, restorePassword);
         restorePasswordTokenRepository.delete(restorePassword);
     }
@@ -49,8 +70,6 @@ public class RestorePasswordService {
         ChangePasswordRequest changePasswordRequest = new ChangePasswordRequest();
         changePasswordRequest.setUser(restorePassword.getUser());
         changePasswordRequest.setPassword(restorePasswordDTO.getNewPassword());
-
         userService.changePassword(changePasswordRequest);
     }
-
 }
